third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-justify-content-004-expected.txt

This is a testharness.js-based test.
FAIL .container > div 1 assert_equals: 
<div class="container" style="justify-content: normal"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 2 assert_equals: 
<div class="container" style="justify-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
FAIL .container > div 3 assert_equals: 
<div class="container" style="justify-content: space-around"><div data-offset-x="6" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 4 assert_equals: 
<div class="container" style="justify-content: space-evenly"><div data-offset-x="6" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 5 assert_equals: 
<div class="container" style="justify-content: stretch"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 6 assert_equals: 
<div class="container" style="justify-content: center"><div data-offset-x="6" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 7 assert_equals: 
<div class="container" style="justify-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
FAIL .container > div 8 assert_equals: 
<div class="container" style="justify-content: end"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 2
FAIL .container > div 9 assert_equals: 
<div class="container" style="justify-content: flex-start"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 10 assert_equals: 
<div class="container" style="justify-content: flex-end"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 11 assert_equals: 
<div class="container" style="justify-content: left"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 10
FAIL .container > div 12 assert_equals: 
<div class="container" style="justify-content: right"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 2
FAIL .container > div 13 assert_equals: 
<div class="container" style="justify-content: normal"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 14 assert_equals: 
<div class="container" style="justify-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
FAIL .container > div 15 assert_equals: 
<div class="container" style="justify-content: space-around"><div data-offset-x="0" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 16 assert_equals: 
<div class="container" style="justify-content: space-evenly"><div data-offset-x="0" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 17 assert_equals: 
<div class="container" style="justify-content: stretch"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 18 assert_equals: 
<div class="container" style="justify-content: center"><div data-offset-x="0" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 19 assert_equals: 
<div class="container" style="justify-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
FAIL .container > div 20 assert_equals: 
<div class="container" style="justify-content: end"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 2
FAIL .container > div 21 assert_equals: 
<div class="container" style="justify-content: flex-start"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 22 assert_equals: 
<div class="container" style="justify-content: flex-end"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 23 assert_equals: 
<div class="container" style="justify-content: left"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got -2
FAIL .container > div 24 assert_equals: 
<div class="container" style="justify-content: right"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 2
Harness: the test ran to completion.
